+++ packages/patrol/darwin/patrol/Package.swift
@@ -62,6 +62,7 @@ let package = Package(
         .process("Resources/de.lproj"),
         .process("Resources/fr.lproj"),
         .process("Resources/pl.lproj"),
+        .process("Resources/ja.lproj"),
       ],
       swiftSettings: patrolImplSwiftSettings,
       linkerSettings: patrolImplLinkerSettings